requirements.txt
transformers==4.40.0
faiss-cpu==1.8.0.post1
numpy==1.26.0
pillow==10.4.0
python==3.12
fitz==0.0.1.dev2
elasticsearch==8.14.0
torch==2.4.0
streamlit==1.37.0
Wikipedia-API==0.6.0
fuzzywuzzy==0.18.0
wikipedia==1.4.0